move 'get timestamp now'
